requirements.txt
tweepy
picamera
pillow
awscli
pimoroni-bme280
smbus